Make tests relying on messageRecorder more reliable.
No it's made sure, that every message taken with getLastMessage can only be taken once.

Reason: The beforehand parameter which was named `timeout` was not really a timeout, but actually a waiting-time to give the response more time to come up (`vopStateChangedNotification` for example may be debounced). The intention was, that when in a command like this
```
triggerActionThenSendsMessageInClient();
const msg = await getLastMessage(driver);
```

`triggerActionThenSendsMessageInClient` takes to long to respond with aa message the getLastMessage could wait a moment and check again. But imagine this scenario:

```
triggerActionThenSendsMessageInClient_1();
const msg1 = await getLastMessage(driver);
triggerActionThenSendsMessageInClient_2();
const msg2 = await getLastMessage(driver);
```
If `triggerActionThenSendsMessageInClient_2` was to slow, `getLastMessage` would return the first message *again*, and do not use the waiting time at all!